rename: diary 폴더 생성
