+++ lib/view_models/diary/diary_camera_view_model.dart
