+++ lib/view_models/diary/diary_record_view_model.dart
