+++ lib/views/diary/diary_post_view.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:nutripic/components/custom_app_bar.dart';
-import 'package:nutripic/view_models/diary_post_view_model.dart';
+import 'package:nutripic/view_models/diary/diary_post_view_model.dart';
 import 'package:provider/provider.dart';
 
 class DiaryPostView extends StatefulWidget {
