+++ lib/views/diary/diary_record_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:nutripic/utils/app_router.dart';
-import 'package:nutripic/view_models/diary_record_view_model.dart';
+import 'package:nutripic/view_models/diary/diary_record_view_model.dart';
 import 'package:provider/provider.dart';
 
 class DiaryRecordView extends StatelessWidget {
